Struct stm32f40x::dma2::hifcr::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn ctcif7(&self) -> CTCIF7R

Bit 27 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif7(&self) -> CHTIF7R

Bit 26 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif7(&self) -> CTEIF7R

Bit 25 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif7(&self) -> CDMEIF7R

Bit 24 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif7(&self) -> CFEIF7R

Bit 22 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif6(&self) -> CTCIF6R

Bit 21 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif6(&self) -> CHTIF6R

Bit 20 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif6(&self) -> CTEIF6R

Bit 19 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif6(&self) -> CDMEIF6R

Bit 18 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif6(&self) -> CFEIF6R

Bit 16 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif5(&self) -> CTCIF5R

Bit 11 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif5(&self) -> CHTIF5R

Bit 10 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif5(&self) -> CTEIF5R

Bit 9 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif5(&self) -> CDMEIF5R

Bit 8 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif5(&self) -> CFEIF5R

Bit 6 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif4(&self) -> CTCIF4R

Bit 5 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif4(&self) -> CHTIF4R

Bit 4 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif4(&self) -> CTEIF4R

Bit 3 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif4(&self) -> CDMEIF4R

Bit 2 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif4(&self) -> CFEIF4R

Bit 0 - Stream x clear FIFO error interrupt flag (x = 7..4)