Struct stm32f40x::dma2::hifcr::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn ctcif7(&mut self) -> _CTCIF7W

Bit 27 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif7(&mut self) -> _CHTIF7W

Bit 26 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif7(&mut self) -> _CTEIF7W

Bit 25 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif7(&mut self) -> _CDMEIF7W

Bit 24 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif7(&mut self) -> _CFEIF7W

Bit 22 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif6(&mut self) -> _CTCIF6W

Bit 21 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif6(&mut self) -> _CHTIF6W

Bit 20 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif6(&mut self) -> _CTEIF6W

Bit 19 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif6(&mut self) -> _CDMEIF6W

Bit 18 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif6(&mut self) -> _CFEIF6W

Bit 16 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif5(&mut self) -> _CTCIF5W

Bit 11 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif5(&mut self) -> _CHTIF5W

Bit 10 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif5(&mut self) -> _CTEIF5W

Bit 9 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif5(&mut self) -> _CDMEIF5W

Bit 8 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif5(&mut self) -> _CFEIF5W

Bit 6 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif4(&mut self) -> _CTCIF4W

Bit 5 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif4(&mut self) -> _CHTIF4W

Bit 4 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif4(&mut self) -> _CTEIF4W

Bit 3 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif4(&mut self) -> _CDMEIF4W

Bit 2 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif4(&mut self) -> _CFEIF4W

Bit 0 - Stream x clear FIFO error interrupt flag (x = 7..4)