Struct stm32f40x::dma2::lifcr::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn ctcif3(&mut self) -> _CTCIF3W

Bit 27 - Stream x clear transfer complete interrupt flag (x = 3..0)

pub fn chtif3(&mut self) -> _CHTIF3W

Bit 26 - Stream x clear half transfer interrupt flag (x = 3..0)

pub fn cteif3(&mut self) -> _CTEIF3W

Bit 25 - Stream x clear transfer error interrupt flag (x = 3..0)

pub fn cdmeif3(&mut self) -> _CDMEIF3W

Bit 24 - Stream x clear direct mode error interrupt flag (x = 3..0)

pub fn cfeif3(&mut self) -> _CFEIF3W

Bit 22 - Stream x clear FIFO error interrupt flag (x = 3..0)

pub fn ctcif2(&mut self) -> _CTCIF2W

Bit 21 - Stream x clear transfer complete interrupt flag (x = 3..0)

pub fn chtif2(&mut self) -> _CHTIF2W

Bit 20 - Stream x clear half transfer interrupt flag (x = 3..0)

pub fn cteif2(&mut self) -> _CTEIF2W

Bit 19 - Stream x clear transfer error interrupt flag (x = 3..0)

pub fn cdmeif2(&mut self) -> _CDMEIF2W

Bit 18 - Stream x clear direct mode error interrupt flag (x = 3..0)

pub fn cfeif2(&mut self) -> _CFEIF2W

Bit 16 - Stream x clear FIFO error interrupt flag (x = 3..0)

pub fn ctcif1(&mut self) -> _CTCIF1W

Bit 11 - Stream x clear transfer complete interrupt flag (x = 3..0)

pub fn chtif1(&mut self) -> _CHTIF1W

Bit 10 - Stream x clear half transfer interrupt flag (x = 3..0)

pub fn cteif1(&mut self) -> _CTEIF1W

Bit 9 - Stream x clear transfer error interrupt flag (x = 3..0)

pub fn cdmeif1(&mut self) -> _CDMEIF1W

Bit 8 - Stream x clear direct mode error interrupt flag (x = 3..0)

pub fn cfeif1(&mut self) -> _CFEIF1W

Bit 6 - Stream x clear FIFO error interrupt flag (x = 3..0)

pub fn ctcif0(&mut self) -> _CTCIF0W

Bit 5 - Stream x clear transfer complete interrupt flag (x = 3..0)

pub fn chtif0(&mut self) -> _CHTIF0W

Bit 4 - Stream x clear half transfer interrupt flag (x = 3..0)

pub fn cteif0(&mut self) -> _CTEIF0W

Bit 3 - Stream x clear transfer error interrupt flag (x = 3..0)

pub fn cdmeif0(&mut self) -> _CDMEIF0W

Bit 2 - Stream x clear direct mode error interrupt flag (x = 3..0)

pub fn cfeif0(&mut self) -> _CFEIF0W

Bit 0 - Stream x clear FIFO error interrupt flag (x = 3..0)