Struct stm32f40x::ethernet_mmc::RegisterBlock

#[repr(C)]

pub struct RegisterBlock {
    pub mmccr: MMCCR,
    pub mmcrir: MMCRIR,
    pub mmctir: MMCTIR,
    pub mmcrimr: MMCRIMR,
    pub mmctimr: MMCTIMR,
    pub mmctgfsccr: MMCTGFSCCR,
    pub mmctgfmsccr: MMCTGFMSCCR,
    pub mmctgfcr: MMCTGFCR,
    pub mmcrfcecr: MMCRFCECR,
    pub mmcrfaecr: MMCRFAECR,
    pub mmcrgufcr: MMCRGUFCR,
    // some fields omitted
}

Register block

Fields
mmccr: MMCCR

0x00 - Ethernet MMC control register

mmcrir: MMCRIR

0x04 - Ethernet MMC receive interrupt register

mmctir: MMCTIR

0x08 - Ethernet MMC transmit interrupt register

mmcrimr: MMCRIMR

0x0c - Ethernet MMC receive interrupt mask register

mmctimr: MMCTIMR

0x10 - Ethernet MMC transmit interrupt mask register

mmctgfsccr: MMCTGFSCCR

0x4c - Ethernet MMC transmitted good frames after a single collision counter

mmctgfmsccr: MMCTGFMSCCR

0x50 - Ethernet MMC transmitted good frames after more than a single collision

mmctgfcr: MMCTGFCR

0x68 - Ethernet MMC transmitted good frames counter register

mmcrfcecr: MMCRFCECR

0x94 - Ethernet MMC received frames with CRC error counter register

mmcrfaecr: MMCRFAECR

0x98 - Ethernet MMC received frames with alignment error counter register

mmcrgufcr: MMCRGUFCR

0xc4 - MMC received good unicast frames counter register