Enum stm32f40x::gpioa::afrl::AFRL7R

pub enum AFRL7R {
    AF0,
    AF1,
    AF2,
    AF3,
    AF4,
    AF5,
    AF6,
    AF7,
    AF8,
    AF9,
    AF10,
    AF11,
    AF12,
    AF13,
    AF14,
    AF15,
}

Possible values of the field AFRL7

Variants
AF0

0000: AF0

AF1

0001: AF1

AF2

0010: AF2

AF3

0011: AF3

AF4

0100: AF4

AF5

0101: AF5

AF6

0110: AF6

AF7

0111: AF7

AF8

1000: AF8

AF9

1001: AF9

AF10

1010: AF10

AF11

1011: AF11

AF12

1100: AF12

AF13

1101: AF13

AF14

1110: AF14

AF15

1111: AF15

Methods

impl AFRL7R

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_af0(&self) -> bool

Checks if the value of the field is AF0

pub fn is_af1(&self) -> bool

Checks if the value of the field is AF1

pub fn is_af2(&self) -> bool

Checks if the value of the field is AF2

pub fn is_af3(&self) -> bool

Checks if the value of the field is AF3

pub fn is_af4(&self) -> bool

Checks if the value of the field is AF4

pub fn is_af5(&self) -> bool

Checks if the value of the field is AF5

pub fn is_af6(&self) -> bool

Checks if the value of the field is AF6

pub fn is_af7(&self) -> bool

Checks if the value of the field is AF7

pub fn is_af8(&self) -> bool

Checks if the value of the field is AF8

pub fn is_af9(&self) -> bool

Checks if the value of the field is AF9

pub fn is_af10(&self) -> bool

Checks if the value of the field is AF10

pub fn is_af11(&self) -> bool

Checks if the value of the field is AF11

pub fn is_af12(&self) -> bool

Checks if the value of the field is AF12

pub fn is_af13(&self) -> bool

Checks if the value of the field is AF13

pub fn is_af14(&self) -> bool

Checks if the value of the field is AF14

pub fn is_af15(&self) -> bool

Checks if the value of the field is AF15