Enum stm32f40x::gpioa::ospeedr::OSPEEDR15R

pub enum OSPEEDR15R {
    LOWSPEED,
    MEDIUMSPEED,
    FASTSPEED,
    HIGHSPEED,
}

Possible values of the field OSPEEDR15

Variants
LOWSPEED

00: Low speed

MEDIUMSPEED

01: Medium speed

FASTSPEED

10: Fast speed

HIGHSPEED

11: High speed

Methods

impl OSPEEDR15R

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_low_speed(&self) -> bool

Checks if the value of the field is LOWSPEED

pub fn is_medium_speed(&self) -> bool

Checks if the value of the field is MEDIUMSPEED

pub fn is_fast_speed(&self) -> bool

Checks if the value of the field is FASTSPEED

pub fn is_high_speed(&self) -> bool

Checks if the value of the field is HIGHSPEED