Enum stm32f40x::gpioa::pupdr::PUPDR15R

pub enum PUPDR15R {
    NOPULLUPPULLDOWN,
    PULLUP,
    PULLDOWN,
}

Possible values of the field PUPDR15

Variants
NOPULLUPPULLDOWN

00: No pull-up, pull-down

PULLUP

01: Pull-up

PULLDOWN

10: Pull-down

Methods

impl PUPDR15R

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_no_pull_up_pull_down(&self) -> bool

Checks if the value of the field is NOPULLUPPULLDOWN

pub fn is_pull_up(&self) -> bool

Checks if the value of the field is PULLUP

pub fn is_pull_down(&self) -> bool

Checks if the value of the field is PULLDOWN