Struct stm32f40x::gpioi::afrh::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn afrh15(&self) -> AFRH15R

Bits 28:31 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh14(&self) -> AFRH14R

Bits 24:27 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh13(&self) -> AFRH13R

Bits 20:23 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh12(&self) -> AFRH12R

Bits 16:19 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh11(&self) -> AFRH11R

Bits 12:15 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh10(&self) -> AFRH10R

Bits 8:11 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh9(&self) -> AFRH9R

Bits 4:7 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh8(&self) -> AFRH8R

Bits 0:3 - Alternate function selection for port x bit y (y = 8..15)