Struct stm32f40x::gpioi::afrl::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn afrl7(&self) -> AFRL7R

Bits 28:31 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl6(&self) -> AFRL6R

Bits 24:27 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl5(&self) -> AFRL5R

Bits 20:23 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl4(&self) -> AFRL4R

Bits 16:19 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl3(&self) -> AFRL3R

Bits 12:15 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl2(&self) -> AFRL2R

Bits 8:11 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl1(&self) -> AFRL1R

Bits 4:7 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl0(&self) -> AFRL0R

Bits 0:3 - Alternate function selection for port x bit y (y = 0..7)