Enum stm32f40x::syscfg::exticr1::EXTI3R

pub enum EXTI3R {
    PA_X,
    PB_X,
    PC_X,
    PD_X,
    PE_X,
    PH_X,
    _Reserved(u8),
}

Possible values of the field EXTI3

Variants
PA_X

0000: PA[x] pin

PB_X

0001: PB[x] pin

PC_X

0010: PC[x] pin

PD_X

0011: PD[x] pin

PE_X

0100: PE[x] pin

PH_X

0111: PH[x] pin

_Reserved(u8)

Reserved

Methods

impl EXTI3R

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_pa_x(&self) -> bool

Checks if the value of the field is PA_X

pub fn is_pb_x(&self) -> bool

Checks if the value of the field is PB_X

pub fn is_pc_x(&self) -> bool

Checks if the value of the field is PC_X

pub fn is_pd_x(&self) -> bool

Checks if the value of the field is PD_X

pub fn is_pe_x(&self) -> bool

Checks if the value of the field is PE_X

pub fn is_ph_x(&self) -> bool

Checks if the value of the field is PH_X