Enum stm32f40x::syscfg::pmc::ADC1DC2W

pub enum ADC1DC2W {
    MAINFLASH,
    SYSTEMFLASH,
}

Values that can be written to the field ADC1DC2

Variants
MAINFLASH

0: No effect

SYSTEMFLASH

1: Refer to AN4073 on how to use this bit. Note: These bits can be set only if the following conditions are met:

ADC clock higher or equal to 30 MHz.

Only one ADCxDC2 bit must be selected if ADC conversions do not start at the same time and the sampling times differ.

These bits must not be set when the ADCDC1 bit is set in PWR_CR register.