Enum stm32f40x::usart6::sr::IDLER

pub enum IDLER {
    NOT_DETECTED,
    DETECTED,
}

Possible values of the field IDLE

Variants
NOT_DETECTED

0: No Idle Line is detected

DETECTED

1: Idle Line is detected

Methods

impl IDLER

pub fn bit_is_clear(&self) -> bool

Returns true if the bit is clear (0)

pub fn bit_is_set(&self) -> bool

Returns true if the bit is set (1)

pub fn bit(&self) -> bool

Value of the field as raw bits

pub fn is_not_detected(&self) -> bool

Checks if the value of the field is NOT_DETECTED

pub fn is_detected(&self) -> bool

Checks if the value of the field is DETECTED